Update Playfair_Cipher_Decode.java
This program will decipher text using the Playfair Cipher. Note that the deciphered text will still have "x"'s in it in between double words and at the end if the original parity was odd. These can simply be ignored when reading the message. Also, some "i"'s in the message may actually stand for "j"'s - another by-product of how the Playfair Cipher works!